kacaaki/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
astroid==2.14.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Deprecated==1.2.13
dill==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-filter==22.1
django-multiselectfield==0.1.12
django-oauth-toolkit==2.2.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.0.0
isort==5.12.0
jwcrypto==1.4.2
lazy-object-proxy==1.9.0
Markdown==3.4.1
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
platformdirs==3.1.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==1.7.1
pylint==2.16.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
